improvements but still bad
