+++ lib/todo.dart
@@ -1,0 +1,6 @@
+class Todo {
+  String text;
+  bool done = false;
+  
+  Todo({required this.text});
+}
